requirements.txt

pyzmq==17.1.2
libconf=2.0.1